@@ -1,7 +1,9 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using racesmiths.Data;
 using racesmiths.Models;
+using racesmiths.Models.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -13,15 +15,22 @@ namespace racesmiths.Controllers
     public class HomeController : Controller
     {
         private readonly ILogger<HomeController> _logger;
+        private readonly ApplicationDbContext _context;
 
-        public HomeController(ILogger<HomeController> logger)
+        public HomeController(ILogger<HomeController> logger,ApplicationDbContext context)
         {
             _logger = logger;
+            _context = context;
         }
         [Authorize]
         public IActionResult Index()
         {
-            return View();
+            InfieldViewModel model = new();
+
+            model.Clubs = _context.Clubs.ToList();
+            model.Events = _context.Event.ToList();
+
+            return View(model);
         }
 
         public IActionResult Privacy()
