@@ -24,9 +24,9 @@ namespace racesmiths.Models
 
 
         public virtual ICollection<Club> Clubs { get; set; } = new HashSet<Club>();
-        //public virtual ICollection<Champ> Champs { get; set; } = new HashSet<Champ>();
-        //public virtual ICollection<Event> Events { get; set; } = new HashSet<Event>();
-        //public virtual ICollection<Race> Races { get; set; } = new HashSet<Race>();
+        public virtual ICollection<Champ> Champs { get; set; } = new HashSet<Champ>();
+        public virtual ICollection<Event> Events { get; set; } = new HashSet<Event>();
+        public virtual ICollection<Race> Races { get; set; } = new HashSet<Race>();
 
     }
 }
