@@ -4,12 +4,23 @@
     ViewData["Title"] = "Details";
 }
 
-<h1>Details</h1>
-<div class="row">
-    <div>
+<div class="row my-center">
+    <div class="col-md-8 upcoming">
+        <div class="upcoming-title">
+            Current Championships<hr />
+        </div>
+        @foreach (var item in Model.Champs)
+            {
+                <div class="">
+                    <a class="btn btn-link upcoming-track" asp-controller="Champs" asp-action="Details" asp-route-id="@item.Id">
+                        <span>@Html.DisplayFor(modelItem => item.ChampName)</span><br />
+                    </a>
+                </div>
+            }
+    </div>
+    <div class="col-md">
+        <h2>Create Championship</h2>
         <form asp-controller="Champs" asp-action="Create" enctype="multipart/form-data">
-            @*@Html.Hidden("ClubId")*@
-            @*<input type="hidden" asp-for="ClubId" />*@
             <input type="hidden" name="ClubId" value="@Model.Id" />
 
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
@@ -33,30 +44,10 @@
                 <label name="Settings" class="control-label">Settings</label>
                 <input name="Settings" class="form-control" />
             </div>
-            @*<div class="form-group">
-                    <label asp-for="ChampType" class="control-label"></label>
-                    <select asp-for="ChampType" class="form-control"></select>
-                    <span asp-validation-for="ChampType" class="text-danger"></span>
-                </div>*@
             <div class="form-group">
                 <label name="StartDate" class="control-label">StartDate</label>
                 <input name="StartDate" class="form-control" />
             </div>
-            @*<div class="form-group">
-                    <label asp-for="Status" class="control-label"></label>
-                    <select asp-for="Status" class="form-control"></select>
-                    <span asp-validation-for="Status" class="text-danger"></span>
-                </div>*@
-            @*<div class="form-group">
-                    <label asp-for="ImagePath" class="control-label"></label>
-                    <input asp-for="ImagePath" class="form-control" />
-                    <span asp-validation-for="ImagePath" class="text-danger"></span>
-                </div>
-                <div class="form-group">
-                    <label asp-for="ImageData" class="control-label"></label>
-                    <input asp-for="ImageData" class="form-control" />
-                    <span asp-validation-for="ImageData" class="text-danger"></span>
-                </div>*@
             <div class="form-group">
                 <input type="submit" value="Create" class="btn btn-primary" />
             </div>
@@ -92,7 +83,7 @@
 
 <div class="row my-center my-div">
     <div class="col-md-3">
-        <a asp-area="" asp-controller="Clubs" asp-action="AssignUsers">
+        <a asp-area="" asp-controller="Clubs" asp-action="AssignUsers" asp-route-id="@Model.Id">
             Assign Users
         </a>
     </div>
