@@ -9,9 +9,7 @@
 <table class="table">
     <thead>
         <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.RSUserId)
-            </th>
+         
             <th>
                 @Html.DisplayNameFor(model => model.RoundNumber)
             </th>
@@ -45,9 +43,7 @@
     <tbody>
 @foreach (var item in Model) {
         <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.RSUserId)
-            </td>
+          
             <td>
                 @Html.DisplayFor(modelItem => item.RoundNumber)
             </td>
