@@ -9,60 +9,42 @@
 
 
 
-<div class="row my-center">
-    <div class="col-md-6 div-back">
-        @*<a href="#">
-            <div class="club-link">
-                CORL
+<div class="row my-center mt-2">
+    <div class="col-md-6 upcoming">
+        <div class="upcoming-title">
+            Active Clubs<hr />
+        </div>
+        @foreach (var item in Model.Clubs)
+        {
+            <div>
+                <a class="btn btn-link upcoming-track" asp-controller="Clubs" asp-action="Details" asp-route-id="@item.Id">
+                    <span>@Html.DisplayFor(modelItem => item.ClubName)</span><br />
+                </a>
             </div>
-        </a>
-        <br />
-        <a href="#">
-            <div class="club-link">
-                Cat Racing
-            </div>
-        </a>
-        <br />
-        <a href="#">
-            <div class="club-link">
-                Tommy Fun
-            </div>
-        </a>
-        <br />
-        <a href="#">
-            <div class="club-link">
-                Fast Bastards
-            </div>
-        </a>*@
-
-
-
-        @*@foreach (var item in Model.Clubs)
-            {
-                <div>
-                    <a class="btn btn-light form-control" asp-controller="Clubs" asp-action="Details" asp-route-id="@item.Id">
-                        <span>@Html.DisplayFor(modelItem => item.ClubName)</span>
-                    </a>
-                </div>
-            }*@
+        }
     </div>
     <div class="col-md upcoming">
         <div class="upcoming-title">
-            Next Race <hr />
+            Upcoming races <hr />
         </div>
-        <div class="float-left upcoming-track">
-            Monaco
-        </div>
-      
-        <div class="float-right my-center">Forecast:</div><br />
-        <span class="float-right upcoming-weather">
-            <img src="~/img/icons8-onedrive-32.png" />
-        </span>
-        <span class="float-right upcoming-weather">
-            <img src="~/img/icons8-rain-cloud-50.png" />
-        </span>
-        <span class="float-right upcoming-weather">
-            <img src="~/img/icons8-clouds-32.png" />
-        </span>
+        @foreach (var item in Model.Events)
+        {
+            <div class="upcoming-track">
+                <a asp-controller="Events" asp-action="Details" asp-route-id="@item.Id">
+                    @Html.DisplayFor(modelItem => item.TrackName)
+                </a>
+            </div><br />
+
+        }
+        @*<div class="float-right my-center">Forecast:</div><br />
+            <span class="float-right upcoming-weather">
+                <img src="~/img/icons8-onedrive-32.png" />
+            </span>
+            <span class="float-right upcoming-weather">
+                <img src="~/img/icons8-rain-cloud-50.png" />
+            </span>
+            <span class="float-right upcoming-weather">
+                <img src="~/img/icons8-clouds-32.png" />
+            </span>*@
     </div>
 </div>
