@@ -10,34 +10,35 @@
     <h4>Race</h4>
     <hr />
     <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Position)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Position)
-        </dd>
-        <dt class = "col-sm-2">
+
+        <dt class="col-sm-2">
             @Html.DisplayNameFor(model => model.RaceNumber)
         </dt>
-        <dd class = "col-sm-10">
+        <dd class="col-sm-10">
             @Html.DisplayFor(model => model.RaceNumber)
         </dd>
-        <dt class = "col-sm-2">
+        <dt class="col-sm-2">
             @Html.DisplayNameFor(model => model.RaceLength)
         </dt>
-        <dd class = "col-sm-10">
+        <dd class="col-sm-10">
             @Html.DisplayFor(model => model.RaceLength)
         </dd>
-        <dt class = "col-sm-2">
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Laps)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.Laps)
+        </dd>
+        <dt class="col-sm-2">
             @Html.DisplayNameFor(model => model.QualifyLength)
         </dt>
-        <dd class = "col-sm-10">
+        <dd class="col-sm-10">
             @Html.DisplayFor(model => model.QualifyLength)
         </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Event)
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Event.Description)
         </dt>
-        <dd class = "col-sm-10">
+        <dd class="col-sm-10">
             @Html.DisplayFor(model => model.Event.Description)
         </dd>
     </dl>
@@ -46,3 +47,28 @@
     <a asp-action="Edit" asp-route-id="@Model.Id">Edit</a> |
     <a asp-action="Index">Back to List</a>
 </div>
+
+
+
+<h1>Race Results:</h1><hr />
+<div class="row my-center">
+    <div class="col-4">
+        <h3>Drivers:</h3>
+    </div>
+    <div class="col-2">
+        <h3>Position</h3>
+    </div>
+</div>
+<div class="row my-center">
+    <div class="col-4">
+        Hatemaker777
+    </div>
+    <div class="col-2">
+        <form>
+            <input class="form-control" />
+        </form>
+    </div>
+</div>
+
+
+
